requirements.txt
asgiref==3.7.2
certifi==2022.12.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-appconf==1.0.5
django-environ==0.9.0
django-filter==22.1
django-imagekit==4.1.0
django-import-export==3.2.0
django-resized==1.0.2
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pilkit==2.0
Pillow==9.4.0
psycopg2==2.9.5
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.4.0
tzdata==2022.6
Unipath==1.1
urllib3==1.26.13
whitenoise==6.3.0
xlrd==2.0.1
xlwt==1.3.0
sendgrid==6.10.0
django-recaptcha==3.0.0
django-model-utils==4.3.1